requirements.txt
pip~=25.0.1
attrs~=25.3.0
PySocks~=1.7.1
requests~=2.32.4
pytest~=8.3.5
allure-python-commons~=2.14.3
typing_extensions~=4.13.2
pluggy~=1.6.0
iniconfig~=2.1.0
numpy~=2.2.4
certifi~=2025.4.26
urllib3~=2.4.0
idna~=3.10
charset-normalizer~=3.4.2
packaging~=25.0